Move blog's components under Blog namespace
